refactor(dursto): split test_helpers into database and registry
