+++ durable-storage/xtask/src/main.rs
@@ -14,9 +14,9 @@ use anyhow::Result;
 use anyhow::bail;
 use clap::Parser;
 use clap::Subcommand;
-use octez_riscv_durable_storage::test_helpers::DatabaseOperation;
-use octez_riscv_durable_storage::test_helpers::database_operations_strategy;
-use octez_riscv_durable_storage::test_helpers::make_database_operations;
+use octez_riscv_durable_storage::test_helpers::database::DatabaseOperation;
+use octez_riscv_durable_storage::test_helpers::database::database_operations_strategy;
+use octez_riscv_durable_storage::test_helpers::database::make_database_operations;
 use proptest::prelude::Strategy;
 use proptest::strategy::ValueTree;
 use proptest::test_runner::TestRunner;
